move LinkStyledItemDelegate into widgets.utils.itemdelegates
